fix MarshalText and relocate NextPlus and NextMinus
