+++ decimal64json.go
@@ -1,8 +1,13 @@
 package decimal
 
+import "encoding/json"
+
+var _ json.Marshaler = Decimal64{}
+var _ json.Unmarshaler = (*Decimal64)(nil)
+
 // MarshalText implements the encoding.TextMarshaler interface.
 func (d Decimal64) MarshalJSON() ([]byte, error) {
-	return d.MarshalText(), nil
+	return d.MarshalText()
 }
 
 // UnmarshalText implements the encoding.TextUnmarshaler interface.
